FIX : Cleanup to code
